requirements.lock.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.lock.txt
antlr4-python3-runtime==4.9.3
    # via omegaconf
audioread==3.0.1
    # via librosa
certifi==2025.4.26
    # via requests
cffi==1.17.1
    # via soundfile
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
chatterbox-tts==0.1.1
    # via -r requirements.in
colorama==0.4.6
    # via tqdm
conformer==0.3.2
    # via chatterbox-tts
decorator==5.2.1
    # via librosa
diffusers==0.29.0
    # via chatterbox-tts
distlib==0.3.9
    # via virtualenv
einops==0.8.1
    # via
    #   conformer
    #   s3tokenizer
filelock==3.18.0
    # via
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
    #   virtualenv
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.32.2
    # via
    #   diffusers
    #   tokenizers
    #   transformers
identify==2.6.12
    # via pre-commit
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via diffusers
jinja2==3.1.6
    # via torch
joblib==1.5.1
    # via
    #   librosa
    #   scikit-learn
lazy-loader==0.4
    # via librosa
librosa==0.10.0
    # via chatterbox-tts
llvmlite==0.44.0
    # via numba
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
msgpack==1.1.0
    # via librosa
networkx==3.5
    # via torch
nodeenv==1.9.1
    # via pre-commit
numba==0.61.2
    # via
    #   librosa
    #   resampy
numpy==1.26.0
    # via
    #   chatterbox-tts
    #   diffusers
    #   librosa
    #   numba
    #   onnx
    #   resampy
    #   s3tokenizer
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
    #   transformers
omegaconf==2.3.0
    # via chatterbox-tts
onnx==1.18.0
    # via s3tokenizer
packaging==25.0
    # via
    #   huggingface-hub
    #   lazy-loader
    #   pooch
    #   transformers
pillow==11.2.1
    # via diffusers
platformdirs==4.3.8
    # via
    #   pooch
    #   virtualenv
pooch==1.8.2
    # via librosa
pre-commit==4.2.0
    # via s3tokenizer
protobuf==6.31.1
    # via onnx
pycparser==2.22
    # via cffi
pyside6==6.9.0
    # via -r requirements.in
pyside6-addons==6.9.0
    # via pyside6
pyside6-essentials==6.9.0
    # via
    #   pyside6
    #   pyside6-addons
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   omegaconf
    #   pre-commit
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   diffusers
    #   huggingface-hub
    #   pooch
    #   transformers
resampy==0.4.3
    # via chatterbox-tts
resemble-perth==1.0.1
    # via chatterbox-tts
s3tokenizer==0.1.7
    # via chatterbox-tts
safetensors==0.5.3
    # via
    #   diffusers
    #   transformers
scikit-learn==1.6.1
    # via librosa
scipy==1.15.3
    # via
    #   librosa
    #   scikit-learn
setuptools==80.9.0
    # via -r requirements.in
shiboken6==6.9.0
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
soundfile==0.13.1
    # via librosa
soxr==0.5.0.post1
    # via librosa
sympy==1.13.1
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.20.3
    # via transformers
torch==2.6.0
    # via
    #   chatterbox-tts
    #   conformer
    #   s3tokenizer
    #   torchaudio
torchaudio==2.6.0
    # via
    #   chatterbox-tts
    #   s3tokenizer
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   s3tokenizer
    #   transformers
transformers==4.46.3
    # via chatterbox-tts
typing-extensions==4.13.2
    # via
    #   huggingface-hub
    #   librosa
    #   onnx
    #   torch
urllib3==2.4.0
    # via requests
virtualenv==20.31.2
    # via pre-commit
zipp==3.22.0
    # via importlib-metadata
